Refactor core-infrastructure
